refactor: convert filenames to camelCase
TypeScript ported music utility files' names are converted to 'camelCase' to stick to conventions.
