+++ src/.prototype/musicutils/ts/@types/keySignature.d.ts
